requirements.txt
# Core data science and ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning Models
xgboost>=1.5.0
tensorflow>=2.8.0
keras>=2.8.0
keras-tuner>=1.1.0

# Time Series Analysis
statsmodels>=0.13.0

# Data Processing and Utilities
joblib>=1.1.0
requests>=2.25.0
python-dotenv>=0.19.0

# Development and Testing
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Production Deployment
gunicorn>=20.1.0
flask>=2.0.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.0.0 